app/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
databricks-sdk==0.40.0
databricks-sql-connector==3.7.0
anthropic==0.44.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
sse-starlette==2.2.1
httpx==0.28.1
psycopg2-binary==2.9.10
